backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.46.0
google-generativeai==0.3.2
groq==0.4.2
pandas==2.1.4 